finished test refactor. Execution time down from 9.5 minutes to 5.1 seconds
