+++ Engine/Items/Markings/Marking.cs
@@ -7,7 +7,7 @@ namespace Engine.Items.Markings
     /// <summary>
     /// A "Marking" is left by an interaction. Includes fingerprints, scuff marks, blood stains, bruises, so on.
     /// </summary>
-    class Marking
+    public class Marking
     {
     }
 }
