+++ Tests/UI/Components/CSIKeyBoardCotrolTests.cs
@@ -1,16 +1,11 @@
-﻿using Engine;
-using Engine.Components.UI;
+﻿using Engine.Components.UI;
 using Engine.UI.Components;
 using Engine.Utilities;
 using GoRogue;
 using Microsoft.Xna.Framework.Input;
 using NUnit.Framework;
 using SadConsole;
-using SadConsole.Controls;
 using SadConsole.Input;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using Tests.Mocks;
 
 namespace Tests.UI.Components
@@ -123,7 +118,7 @@ namespace Tests.UI.Components
             Assert.AreEqual(MockGame.Menu, Global.CurrentScreen);
         }
 
-        //[Theory] //todo - just takes too long for now
+        [Theory] //todo - just takes too long for now
         public void QueriableActionOpensACursorTest((GameAction actionkey, GameAction purpose) dataset)
         {
             _lookingGlass = (MagnifyingGlassComponent)_game.Player.GetComponent<MagnifyingGlassComponent>();
@@ -136,13 +131,13 @@ namespace Tests.UI.Components
                 Assert.AreNotEqual(dataset.purpose, _lookingGlass.Purpose);
         }
 
-        //[Theory]//todo -- just takes too long for now
+        [Theory]//todo -- just takes too long for now
         public void TogglesWindowsTest((GameAction action, string windowTitle) dataset)
         {
             Assert.Fail();
         }
 
-        //[Theory]//todo - just takes too long for now
+        [Theory]//todo - just takes too long for now
         public void TakeActionsTest(GameAction action)
         {
             Assert.DoesNotThrow(() => _component.TakeAction(action));
